--- src/main/java/com/grow/member_service/member/application/event/LoginSucceededEvent.java
@@ -1,7 +1,0 @@
-package com.grow.member_service.member.application.event;
-
-import java.time.LocalDateTime;
-
-public record LoginSucceededEvent(Long memberId, LocalDateTime occurredAt) {
-
-}
